health-check-service/requirements.txt
requests
kafka-python
flask
prometheus-flask-exporter